requirements.txt
# Core dependencies for Live Crypto Trading System
websockets>=12.0
aiohttp>=3.9.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
